[fold] router_base is an implementation detail
